validata-table-main/pyproject.toml
[tool.poetry]
name = "validata-table"
version = "0.10.1"
authors = [
  "Amélie Rondot <amelie.rondot@multi.coop>",
]
description = "Package used in validation tabular data"
readme = "README.md"
classifiers = [
        "Development Status :: 5 - Production/Stable",
        "Intended Audience :: Developers",
        "Topic :: Software Development :: Libraries :: Python Modules",
        "Operating System :: OS Independent",
        "License :: OSI Approved :: GNU Affero General Public License v3",
        "Programming Language :: Python :: 3",
]
license = "GPL-3.0-or-later"
repository = "https://gitlab.com/validata-table/validata-table"
packages = [
  { include = "validata_core", from = "src" },
  { include = "validata_api", from = "src" },
  { include = "validata_ui", from = "src" },
]


[tool.poetry.dependencies]
python = "^3.9"
# for validata-core
frictionless = {version = "4.38", extras = ["excel"]}
# for validata-core custom_checks
shapely = "^2.0.1"
python-stdnum = "^1.19"
opening-hours-py = "^0.6.17"
phonenumbers = "^8.13.22"
# for validata-ui
opendataschema = "^0.6.0"
requests-cache = "^1.1.0"
commonmark = "^0.9.1"
pydantic = "1.10.9"
toml = "^0.10.2"
python-dotenv = "^1.0.0"
flask = "2.1.2"
werkzeug = "2.0.2"
jinja2 = "3.0.3"
sentry-sdk = "^1.31.0"
# for validata-api
flasgger = "^0.9.7.1"
flask-matomo = "^1.2.0"
flask-cors = "^4.0.0"
ujson = "^5.8.0"
toolz = "^0.12.0"
cachetools = "^5.3.1"
importlib-metadata = "^6.8.0"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
# for tests
pytest = "^7.4.2"


[tool.poetry.scripts]
validata = 'validata_core.cli:cli'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
